Refactor env conf, read keymaps from file
