+++ config.go
@@ -16,76 +16,21 @@
  **/
 package main
 
-import (
-	"fmt"
-	"os"
-	"strings"
-	"time"
-)
+// Keymaps represents the key mappings in the TOML file
+type Keymaps struct {
+	PrevHour   []string
+	NextHour   []string
+	PrevDay    []string
+	NextDay    []string
+	PrevWeek   []string
+	NextWeek   []string
+	ToggleDate []string
+	OpenWeb    []string
+	Now        []string
+}
 
 // Config stores app configuration
 type Config struct {
-	Zones []*Zone
-}
-
-// LoadConfig from environment
-func LoadConfig(tzConfigs []string) (*Config, error) {
-	conf := Config{
-		Zones: DefaultZones,
-	}
-
-	if len(tzConfigs) == 0 {
-		tzList := os.Getenv("TZ_LIST")
-		if tzList == "" {
-			return &conf, nil
-		}
-		tzConfigs = strings.Split(tzList, ";")
-		if len(tzConfigs) == 0 {
-			return &conf, nil
-		}
-	}
-	zones := make([]*Zone, len(tzConfigs)+1)
-
-	// Setup with Local time zone
-	localZoneName, _ := time.Now().Zone()
-	zones[0] = &Zone{
-		Name:   fmt.Sprintf("(%s) Local", localZoneName),
-		DbName: localZoneName,
-	}
-
-	// Add zones from TZ_LIST
-	for i, zoneConf := range tzConfigs {
-		zone, err := SetupZone(time.Now(), zoneConf)
-		if err != nil {
-			return nil, err
-		}
-		zones[i+1] = zone
-	}
-	conf.Zones = zones
-
-	return &conf, nil
-}
-
-// SetupZone from current time and a zoneConf string
-func SetupZone(now time.Time, zoneConf string) (*Zone, error) {
-	names := strings.Split(zoneConf, ",")
-	dbName := strings.Trim(names[0], " ")
-	var name string
-	if len(names) == 2 {
-		name = names[1]
-	}
-
-	loc, err := time.LoadLocation(dbName)
-	if err != nil {
-		return nil, fmt.Errorf("looking up zone %s: %w", dbName, err)
-	}
-	if name == "" {
-		name = loc.String()
-	}
-	then := now.In(loc)
-	shortName, _ := then.Zone()
-	return &Zone{
-		DbName: loc.String(),
-		Name:   fmt.Sprintf("(%s) %s", shortName, name),
-	}, nil
+	Zones   []*Zone
+	Keymaps Keymaps
 }
