+++ example-conf.toml
@@ -17,13 +17,12 @@ id = "UTC"
 name = "UTC"
 
 [keymaps]
-prev_hour = ["h"]
-next_hour = ["l"]
-prev_day = ["k"]
-next_day = ["j"]
+prev_hour = ["h", "left"]
+next_hour = ["l", "right"]
+prev_day = ["k", "up"]
+next_day = ["j", "down"]
 prev_week = ["p"]
 next_week = ["n"]
 toggle_date = ["d"]
 open_web = ["o"]
 now = ["t"]
-
